Settings moved to User Control
Settings carefully moved to user control, saving and restoring them.
Seems to work - needs testing. Tested adding a commander, the tick boxes
work, the log loc works.
Now to update theme more
