+++ EDDiscovery/EDDiscoveryForm.resx
@@ -120,18 +120,6 @@
   <metadata name="folderBrowserDialog1.TrayLocation" type="System.Drawing.Point, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a">
     <value>17, 17</value>
   </metadata>
-  <metadata name="ColumnNr.UserAddedColumn" type="System.Boolean, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089">
-    <value>True</value>
-  </metadata>
-  <metadata name="ColumnCommander.UserAddedColumn" type="System.Boolean, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089">
-    <value>True</value>
-  </metadata>
-  <metadata name="ColumnEDSMAPIKey.UserAddedColumn" type="System.Boolean, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089">
-    <value>True</value>
-  </metadata>
-  <metadata name="ColumnNetLogPath.UserAddedColumn" type="System.Boolean, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089">
-    <value>True</value>
-  </metadata>
   <metadata name="menuStrip1.TrayLocation" type="System.Drawing.Point, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a">
     <value>186, 17</value>
   </metadata>
@@ -7666,15 +7654,11 @@
         AABJRU5ErkJggg==
 </value>
   </data>
-  <data name="panel_close.BackgroundImage" type="System.Drawing.Bitmap, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
+  <data name="panel_grip.BackgroundImage" type="System.Drawing.Bitmap, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
     <value>
-        iVBORw0KGgoAAAANSUhEUgAAABgAAAAYCAYAAADgdz34AAAABGdBTUEAALGPC/xhBQAAAAlwSFlzAAAO
-        vAAADrwBlbxySQAAAStJREFUSEvtU83KgkAUFaqFIvgvbhXdKk7UO5RWz1j0ilmr+s75uBPSKgfaeWCY
-        mev1nPs31owZv0cYhgvXdddyfQO2TRRFS7maIU3TZRAE5zzPB5D1YqZoXxTFgG/XLMtWYp4Gz/NIfqnr
-        +qmUepVlOcRx3EGo45m2pmmeFIHg9Exs21aI/Na27b8AdxA/qqp68KztzATl2spv04Do9iC9aQG96wWx
-        AdGfxN0MINhB5P4pwExQsqO4mQM1P5KM5ONFUQjsxM0MTJ9l0JHrXS+WDz57cZ8GNHnLBo6jZiZs9NjG
-        QUCTlfz2PXzf55heOYokAjEb2mH1PDMDjjBHGTazB5ckyYoizITEYuZ0Hfj4sJ/5GMVsBmbiOM5Grm/A
-        tsZjXMh1xoyfwbL+AJ2UdD7DHVmcAAAAAElFTkSuQmCC
+        iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAABGdBTUEAALGPC/xhBQAAAAlwSFlzAAAe
+        vAAAHrwB7kRN9wAAAAd0SU1FB+ADEQ4BMEDJ1+8AAAA0SURBVDhPY6AQ/IfSpIEFCxaANI5qBgKoOH5A
+        QDN+A0Y1IwABcQggUzNEfOA0QwEFmhkYAAqwdAUciUDwAAAAAElFTkSuQmCC
 </value>
   </data>
   <data name="panel_minimize.BackgroundImage" type="System.Drawing.Bitmap, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
@@ -7684,11 +7668,15 @@
         D/7NaCIi8ohfduQARsU2X3DluFNs8wVXjjvFNl9w+3Ca/MVsARL/eYd5ibczAAAAAElFTkSuQmCC
 </value>
   </data>
-  <data name="panel_grip.BackgroundImage" type="System.Drawing.Bitmap, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
+  <data name="panel_close.BackgroundImage" type="System.Drawing.Bitmap, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
     <value>
-        iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAABGdBTUEAALGPC/xhBQAAAAlwSFlzAAAe
-        vAAAHrwB7kRN9wAAAAd0SU1FB+ADEQ4BMEDJ1+8AAAA0SURBVDhPY6AQ/IfSpIEFCxaANI5qBgKoOH5A
-        QDN+A0Y1IwABcQggUzNEfOA0QwEFmhkYAAqwdAUciUDwAAAAAElFTkSuQmCC
+        iVBORw0KGgoAAAANSUhEUgAAABgAAAAYCAYAAADgdz34AAAABGdBTUEAALGPC/xhBQAAAAlwSFlzAAAO
+        vAAADrwBlbxySQAAAStJREFUSEvtU83KgkAUFaqFIvgvbhXdKk7UO5RWz1j0ilmr+s75uBPSKgfaeWCY
+        mev1nPs31owZv0cYhgvXdddyfQO2TRRFS7maIU3TZRAE5zzPB5D1YqZoXxTFgG/XLMtWYp4Gz/NIfqnr
+        +qmUepVlOcRx3EGo45m2pmmeFIHg9Exs21aI/Na27b8AdxA/qqp68KztzATl2spv04Do9iC9aQG96wWx
+        AdGfxN0MINhB5P4pwExQsqO4mQM1P5KM5ONFUQjsxM0MTJ9l0JHrXS+WDz57cZ8GNHnLBo6jZiZs9NjG
+        QUCTlfz2PXzf55heOYokAjEb2mH1PDMDjjBHGTazB5ckyYoizITEYuZ0Hfj4sJ/5GMVsBmbiOM5Grm/A
+        tsZjXMh1xoyfwbL+AJ2UdD7DHVmcAAAAAElFTkSuQmCC
 </value>
   </data>
   <metadata name="$this.TrayHeight" type="System.Int32, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089">
